[back] add chat-service - split the gateway and the service
